+++ opensquirrel/passes/decomposer/__init__.py
@@ -8,6 +8,7 @@ from opensquirrel.passes.decomposer.aba_decomposer import (
 )
 from opensquirrel.passes.decomposer.cnot2cz_decomposer import CNOT2CZDecomposer
 from opensquirrel.passes.decomposer.cnot_decomposer import CNOTDecomposer
+from opensquirrel.passes.decomposer.cz_decomposer import CZDecomposer
 from opensquirrel.passes.decomposer.general_decomposer import Decomposer
 from opensquirrel.passes.decomposer.mckay_decomposer import McKayDecomposer
 from opensquirrel.passes.decomposer.swap2cnot_decomposer import SWAP2CNOTDecomposer
@@ -16,6 +17,7 @@ from opensquirrel.passes.decomposer.swap2cz_decomposer import SWAP2CZDecomposer
 __all__ = [
     "CNOT2CZDecomposer",
     "CNOTDecomposer",
+    "CZDecomposer",
     "Decomposer",
     "McKayDecomposer",
     "SWAP2CNOTDecomposer",